# openapi.yaml
openapi: 3.1.0
info:
  title: Bakame API
  version: 1.0.0
  description: |
    Bakame - Production-grade AI platform API

    All endpoints except /health require Bearer token authentication.

    Response Format:
    - Success: { data: <payload>, meta: { requestId: string } }
    - Error: { error: { code: string, message: string, requestId: string } }
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'

    patch:
      tags: [Users]
      summary: Update profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  nullable: true
                avatarUrl:
                  type: string
                  nullable: true
                timezone:
                  type: string
                locale:
                  type: string
      responses:
        '200':
          description: Profile updated
  /users/me/preferences:
    get:
      tags: [Users]
      summary: Get AI preferences
      operationId: getPreferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: AI preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'

    put:
      tags: [Users]
      summary: Update AI preferences
      operationId: updatePreferences
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                responseLength:
                  type: string
                  enum: [concise, balanced, detailed]
                formality:
                  type: string
                  enum: [casual, neutral, formal]
                allowMemory:
                  type: boolean
                allowWebSearch:
                  type: boolean
                customInstructions:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Preferences updated
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - PERMISSION_DENIED
                - NOT_FOUND
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
            requestId:
              type: string
...
    ProfileResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            userId:
              type: string
            displayName:
              type: string
              nullable: true
            avatarUrl:
              type: string
              nullable: true
            timezone:
              type: string
            locale:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
    PreferencesResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            userId:
              type: string
            responseLength:
              type: string
              enum: [concise, balanced, detailed]
            formality:
              type: string
              enum: [casual, neutral, formal]
            allowMemory:
              type: boolean
            allowWebSearch:
              type: boolean
            customInstructions:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string